delete and edit authorization added

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const {validationResult} = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -47,20 +48,28 @@ exports.createPost = (req, res, next) => {
       throw error;
     }
     const imageUrl = req.file.path.replace("\\" ,"/");
-    const {title,content} = req.body
+    const {title,content} = req.body;
+    let creator;
+
     const post = new Post({
       title, 
       content,
       imageUrl,
-      creator: {name: 'Sehan'}
+      creator: req.userId //fetched from the middleware
     })
     post.save()
     .then(result => {
-      console.log(result);
-      // Create post in db
-      res.status(201).json({
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post); //mongoose will extract the id and populate
+      return user.save();
+    })
+    .then(result => {res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator: {_id: creator._id, name: creator.name}
       });
     })
     .catch(err => {
@@ -123,6 +132,12 @@ exports.updatePost = (req,res,next) => {
       throw error;
     }
 
+    if(post.creator.toString() !== req.userId){
+      const error = new Error('Not authorized');
+      error.statusCode = 403;
+      throw error;
+    }
+
     if (imageUrl !== post.imageUrl){
       clearImage(post.imageUrl);
     }
@@ -157,12 +172,23 @@ exports.deletePost = (req,res,next) => {
       throw error;
     }
 
-    //TODO: check logged in user is the owner ?
+    if(post.creator.toString() !== req.userId){
+      const error = new Error('Not authorized');
+      error.statusCode = 403;
+      throw error;
+    }
+
     clearImage(post.imageUrl);
     return Post.findByIdAndRemove(postId);
   })
   .then(result => {
-    console.log(result);
+    return User.findById(req.userId);
+  })
+  .then(user => {
+    user.posts.pull(postId);
+    return user.save()
+  })
+  .then(result => {
     res.status(200).json({
       message: 'Deleted Post!'
     });
